Widget Initialization
In this version, the initialization procedures of QVTKWidget, TextEdit, and four kinds of layout are finished. Moreover, add the 3rd Party Library configuration including PCL 1.8.0 and VTK 7.0.0.

But there is a bug that when change the from layout 4 to layout 2, from layout 4 to layout 2 and from layout 2 to layout 1, there is a unknown shadow remained.