Deal with Backspace for TextEdit_CMD
This version of the coding deals with the event filter for the TextEdit_CMD.

In order to make a command line interactive mode, TextEdit_CMD should perform like a CMD where the record cannot be deleted without operations and can accept the new instruction at the same time.

There are two kinds of ways to delete the characters: one at once and delete the selected text.