Improve TextEdit_CMD
Improve the TextEdit_CMD including its eventfilter that deals with the delete key events and other normal keys (number keyboard excluded).

Add codes dealing with the color of input and output in white and grey.

There is a bug remained that if there is selected text and number keyboard event happened at the same time, the eventfilter cannot deal with the number keyboard events. It seems like number keyboard is not within the same keyevent system as other normal keys.