Deal with Enter for TextEdit_CMD
Add eventfilter for Enter in TextEdit_CMD for command process. In order to process the command, a Command Analyser and a Command Processor are needed, where the Command Analyser is used for detecting if the input is a command or not, and Command Processor is used for processing the Command as a thread.