Design Command System
After getting the input from TextEdit_CMd, there should be a system deal with the input.

Here I pack up the input into a Command system in order to make the communication between main thread and subthread neat and organized.

The command can be divided into three parts including predicate, object, and complement to tell the main thread and subthread do what kind of the work. If there is a need to add more command, just add the keywords into the dictionary QStringList.