Finish the thread communication
After sending the command to the Command Processor, the Command Processor deal with command and make reaction.

I design the Command Precessor as a Qthread, so it's better to do the thread communication between the main thread and Command Precessor.

Using the Slot and Signal techniques, the inter thread communication is becomming possible.