Finished Layout Change Command
There was a bug when processing the Layout Change Command. When the program try to process the Layout Change Command, the previous QVTKWidget cannot be removed.

There isn't a good way to clear the widget, so I clear the content in QVTKWidget part in order to have a clean background.