Fix the bug after resetting the camera
There is a bug after changing the layout and synchronizing the camera continuiously, the Widgets will change the size of its own.

After changing the update code from viewer->updateCamera() to qvtkWidget->update(). The bug is fixed.

But the bug remains that when change from the layout_6 to layout_2, widget size will change too.

What should be finished next is to finish functions related to layout 6 and figure out the reason leading to the bug discussed above.